Rename: category.ts -> categoryDto.ts
